+++ docs/CHANGELOG.md
@@ -1,4 +1,10 @@
-Version: v0.1.1
+Version: v0.1.2
 
 Commit:
 feat: ChampionData自動生成スクリプトを追加
+
+## v0.1.2
+
+### Added
+- ChampionData Editor
+- champions.jsonからgenerated.tsを生成可能